RSC convert admin referral-program page
